Removed IO support, improved parser and cpu
IO support was experimental and hacky, and it turns out it didn't work
as supposed to. apparently there's some magic involved in statements
like:
ADD R0 IO IO. If IO is 10 and IO is 5 it adds them and puts them in R0.
At the moment I don't know how to emulate that.